14/03/13
Continued to work on the playercontroller class, changed the
dynamicobject class to work with the controller and implemented a
decorator pattern
for it.

+++ MMO_Game/Structs.h
@@ -24,7 +24,7 @@ struct Physics{
 	Ogre::Vector3 _Facing;
 
 	int _MovementSpeed;
-	int _RotationSpeed;
+	float _RotationSpeed;
 
 	float _Mass;
